+++ ls-plugin.php
@@ -44,6 +44,7 @@ add_action( 'init', 'ls_plugin_load_textdomain' );
  * Add include files to inc/ and require them here when ready.
  */
 function ls_plugin_init() {
+	require_once LS_PLUGIN_PLUGIN_DIR . 'inc/linkable-blocks.php';
 	require_once LS_PLUGIN_PLUGIN_DIR . 'inc/class-style-switcher.php';
 
 	$style_switcher = new LS_Plugin_Style_Switcher();